+++ changelog.md
@@ -78,6 +78,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `curvlinops.examples.gradient_and_loss` as the replacement utility function
   ([PR](https://github.com/f-dangel/curvlinops/pull/245))
 
+- Make `_ChainPyTorchLinearOperator` support more than two operators
+  ([PR](https://github.com/f-dangel/curvlinops/pull/246))
+
 ## [3.0.1] - 2026-01-14
 
 This patch provides major performance improvements for all curvature matrices (by using `torch.func`, and by improving EKFAC's eigenvalue correction) and deprecates Python 3.9 (we now require at least 3.10).